dwm.c: Formatting + scrolling title + constant bar updating.
The currently focused window's title now slowly scrolls to the right of
the screen until it hits the widgets section, at which point it will go
back at the beginning.

In order to implement the aforementioned feature we need to update the
current bar with drawbar() constantly. Right now this is done 60
times per second.