widget: Add some widget helper functions + async widgets.
Add 4 helper functions, 2 of which help with the initialization and the
updating of widgets.

The other two have to do with asynchronous widgets. See below.

We can now have widgets that get updated from another thread in a safe
manner. This is done by setting Widget::_should_lock_on_access to true,
and initializing the Widget::_lock using mtx_init(). When a widget is
getting updated from another thread it should hold a lock on itself by
either using widget_lock()/widget_unlock() or mtx_lock()/mtx_unlock().

Similarly when the main dwm thread accesses widgets, it holds a lock
during it's access time.

This was implemented mainly for a future PulseAudio client connection.
As much as I would have liked to keep everything single-threaded the
PulseAudio C API pretty much forces us to spawn a separate thread if we
want to do anything interesting.